tools/esp32_tool/requirements.txt
adafruit_board_toolkit==1.1.1
esptool==4.6.2
pyserial==3.5
PySide2==5.15.2.1
